pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-meta_hackathon"
version = "0.1.0"
description = "Meta Hackathon environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    # install from github
    # "openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git",
    "openenv-core[core]>=0.2.2",
    # Environment-specific dependencies
    # Add all dependencies needed for your environment here
    # Examples:
    # "numpy>=1.19.0",
    # "torch>=2.0.0",
    # "gymnasium>=0.29.0",
    # "openspiel>=1.0.0",
    # "smolagents>=1.22.0,<2",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.31.0",
    "openai>=1.30.0",
    "pyyaml>=6.0.2",
    "websockets>=12.0",
    "uvicorn[standard]>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m meta_hackathon.server.app
server = "meta_hackathon.server.app:main"
inference = "meta_hackathon.inference:main"

[tool.setuptools]
include-package-data = true
packages = [
    "meta_hackathon",
    "meta_hackathon.server",
    "meta_hackathon.agent",
    "meta_hackathon.cicd",
]

[tool.setuptools.package-dir]
meta_hackathon = "."
"meta_hackathon.server" = "server"
"meta_hackathon.agent" = "agent"
"meta_hackathon.cicd" = "cicd"
